Moved locales directories up one level.
