+++ packages/react-components/src/locales/de-DE/strings.json
