+++ packages/react-components/src/locales/en-GB/strings.json
