+++ packages/react-components/src/locales/es-ES/strings.json
