+++ packages/react-components/src/locales/fr-FR/strings.json
