+++ packages/react-components/src/locales/index.ts
