+++ packages/react-components/src/locales/ja-JP/strings.json
