+++ packages/react-components/src/locales/ko-KR/strings.json
