+++ packages/react-components/src/locales/pt-BR/strings.json
