+++ packages/react-components/src/locales/ru-RU/strings.json
