+++ packages/react-components/src/locales/zh-CN/strings.json
