+++ packages/react-composites/src/composites/index.ts
@@ -8,6 +8,6 @@ export * from './CallComposite';
 export * from './MeetingComposite';
 export type { AvatarPersonaData, AvatarPersonaDataCallback } from './common/AvatarPersona';
 export * from './common/icons';
-export * from './localization/locales';
+export * from './locales';
 export type { CompositeStrings, CompositeLocale } from './localization';
 export type { BaseCompositeProps } from './common/Composite';
