+++ packages/react-composites/src/composites/locales/de-DE/strings.json
