+++ packages/react-composites/src/composites/locales/es-ES/strings.json
