+++ packages/react-composites/src/composites/locales/fr-FR/strings.json
