+++ packages/react-composites/src/composites/locales/ja-JP/strings.json
