+++ packages/react-composites/src/composites/locales/ko-KR/strings.json
