+++ packages/react-composites/src/composites/locales/nl-NL/strings.json
