+++ packages/react-composites/src/composites/locales/pt-BR/strings.json
