+++ packages/react-composites/src/composites/locales/zh-CN/strings.json
